Initial commit
Create a simplified build system to build various third party packages
as external projects. This is loosely based on Fletch¹, but has been
significantly stripped down, as it is intended to only work (for now) on
Ubuntu 18.04, and only in a docker container being used to build the
Drake wheel for PyPI. Accordingly, we can make quite a few assumptions
about the build environment.

At some point, we may add macOS support, but most of the weird logic is
for handling Windows.

This isn't complete, but it appears to cover most of the critical
dependencies, with the exception of glib2 and possibly glew ("possibly"
in the sense that I'm hoping we'll either be able to drop that one from
the critical set or use a packaged version).

(¹ https://github.com/Kitware/fletch/)