Add "stubs" for CLP, IPOPT; remove VTK
Add stubs for clp and ipopt which just generate "empty" pkg-config
files. Drake fails hard if these don't exist, even when disabling their
use; this should get us past that for now. (Eventually we probably want
the real things, but for now we're just trying to get some minimal
product built as soon as possible.)

Also remove VTK; due to how drake's build consumes it, we're going to
need to take a different approach that will exclude that being build via
this repository. Also remove glew, which seems to only be used by VTK,
and because VTK provides an internal version that we should be able to
use.