Add VTK
Add a script to build VTK, and execute this as an additional step.
Separating this from the other dependencies is potentially useful, as we
may need to distribute VTK differently, and also because it takes much
longer to build than everything else.

Also, tweak things so that the scripts are run directly rather than
invoking the shell explicitly. This is needed for the #! lines (and
crucially, their use of -e) to actually have an effect.