Disable shared libraries for dependencies
Modify the configure/build rules for dependencies which by default build
both shared and static libraries to only build the static libraries...
for all such dependencies except zlib, which does not have any way of
disabling the shared library.

This will help ensure that the shared libraries are not used. We may
need to either patch zlib or forcibly nuke its .so.