Rename Drake build script
Rename Drake build script, which doesn't actually generate the wheel, to
free up the name for a separate script that does. This will allow the
"install drake" and "create wheel" steps to be separated in the Docker
build, allowing the former to be cached while the latter is under active
development.

This separation may not be permanent, however, since eventually building
the wheel should probably move to bazel.

There are otherwise no functional changes; this is being done as a
separate commit so git can track the script through the rename (since
adding a "new" file with an existing name breaks rename detection).