Build our own patchelf
Update provisioning and dependencies to build a newer version of
patchelf rather than using the version on Bionic, which apparently does
Bad Things as used by auditwheel, resulting in a libdrake.so that causes
the loader to abort with an assertion failure.

Note that we are using 0.12, as 0.13 SEGVs during the Drake build.