Use virtualenv Python and "disable" debugging
Replace both the standard and debugging Python interpreters with the
standard interpreter from our virtualenv. This is correct for the
former, and for the latter, should suffice to make the build system
happy, while not ever actually being used during a wheel build (so it
won't matter that it isn't a debug interpreter).

This should be sufficient that Drake can build. However, due to some
nasty business with libcrypt, which for some reason is not used by
Python 3.6, but is required by Python 3.7, we are not yet able to
produce a PyPI-usable wheel.

See also https://github.com/pypa/manylinux/issues/305 for an explanation
of why this is broken.

fixup