Remove much of pip-drake.patch
With e637144d8deadeabd2c99638d9af2d34ffbc5f8c landed in upstream Drake,
most of the monkey-patching we were doing to get Drake to build in the
Docker wheel building environment are no longer needed; these can (and
indeed must, due to conflicts) be removed.

The remaining LCM patch is very much a temporary hack that will also
need to be removed eventually. The remaining Python patch may or may not
be permanent.